refactor ml api package to set scala version as default
